requirements.txt
streamlit
python-dotenv
requests
crewai
pydantic
markitdown